Fixed observer collection tests + some minor cleanup

+++ Source/Orleankka.Tests/Scenarios/Using_reminders.cs
@@ -15,13 +15,13 @@ namespace Orleankka.Scenarios
         public async void When_reminder_is_fired_an_instance_of_correct_actor_type_should_be_activated()
         {
             var actor = system.FreshActorOf<TestActor>();
-            var hashcode = await actor.Ask<int>(new GetInstanceHashcode());
+            var hashcode = await actor.Ask<long>(new GetInstanceHashcode());
             
             await actor.Tell(new SetReminder());
             await Task.Delay(TimeSpan.FromSeconds(90));
 
             Assert.True(await actor.Ask<bool>(new HasBeenReminded()));
-            Assert.AreNotEqual(hashcode, await actor.Ask<int>(new GetInstanceHashcode()));
+            Assert.AreNotEqual(hashcode, await actor.Ask<long>(new GetInstanceHashcode()));
         } 
     }
 }
